+++ crate/input_model/src/thing.rs
@@ -1,11 +1,12 @@
 pub use self::{
-    thing_copy_text::ThingCopyText, thing_dependencies::ThingDependencies,
+    thing_copy_text::ThingCopyText, thing_dependencies::ThingDependencies, thing_descs::ThingDescs,
     thing_hierarchy::ThingHierarchy, thing_id::ThingId, thing_interactions::ThingInteractions,
     thing_layouts::ThingLayouts, thing_names::ThingNames,
 };
 
 mod thing_copy_text;
 mod thing_dependencies;
+mod thing_descs;
 mod thing_hierarchy;
 mod thing_id;
 mod thing_interactions;