+++ crate/ir_model/src/entity.rs
@@ -1,4 +1,4 @@
-pub use disposition_model_common::entity::{EntityDescs, EntityType, EntityTypeId, EntityTypes};
+pub use disposition_model_common::entity::{EntityType, EntityTypeId, EntityTypes};
 
 pub use self::entity_tailwind_classes::EntityTailwindClasses;