+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 * Add `EdgeLabel` to allow different labels at each end of an edge. ([#32][#32])
 * Render edge descriptions in the middle of the edge path. ([#32][#32])
+* Replace `EntityDescs` with `ThingDescs` and `EdgeDescs`. ([#33][#33])
 
 [#32]: https://github.com/azriel91/disposition/pull/32
+[#33]: https://github.com/azriel91/disposition/pull/33
 
 
 ## 0.2.0 (2026-05-22)